feat (makefile): Creates linting/formatting shortcuts + Compiler control
This commit creates shortcuts that will call upon clang-tidy and
clang-format for linting and formatting purposes. This allows for easier
and more abstracted usage for other programmers and CI/CD

This will needed to be added to programmer-facing documentation.

Additionally, changes to common.mk include controls for lower level
debugging control to the compiler, with a toggle set for setting on and
off compiler flags, including:

- Compiler static analysis
- Compiler optimization flags
- Compiler sanitization flags