chore (constants+integration): Integrate
The custom controller class allows the partner controller to send notifications
 to the main controller. It also allows the main controller to be switched
between each from a keypress from the previous main controller.

This specific commit integrates the custom controller class into the
main.cpp file - allowing it to be used in driver control, which is always
unless the controller is connected to a competition switch.